requirements.txt
# UI Framework
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.0.0
python-dateutil==2.8.2

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0  # PostgreSQL driver
aiosqlite>=0.19.0  # Async SQLite support

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
cryptography>=41.0.0
pyjwt>=2.8.0

# LLM Integration
requests>=2.31.0
aiohttp>=3.9.0
openai>=1.0.0  # For OpenAI-compatible APIs
tiktoken>=0.5.0  # Token counting

# Environment & Configuration
python-dotenv>=1.0.0

# Logging & Monitoring
python-json-logger>=2.0.0
sentry-sdk>=1.38.0

# Caching (optional)
redis>=5.0.0
hiredis>=2.2.0

# Email (optional)
python-email-validator>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0  # For testing FastAPI

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
pylint>=3.0.0

# Utilities
click>=8.1.0
pyyaml>=6.0.0
typing-extensions>=4.8.0
